Delete unnecessary check for IEnum of Object
The ParameterManager method public String AddParameter( Object
parameterValue ) was checking the parameterValue for both
IEnumerable<Object> and IEnumerable (non-generic). The non-generic check
is all that is necessary. Additionally, the non-generic check works for
the 2.0 CLR. The generic-based IEnumerable<Object> depended on the 4.0
CLR to recognize that an IEumerable<String> was an IEnumerable<Object>.
This led to runtime errors in the 2.0 CLR.